pyproject.toml
[tool.poetry]
name = "virtool"
version = "0.0.0"
authors = [
    "Ian Boyes",
    "Reece Hoffmann",
    "Aman Monga",
    "Blake Smith",
    "Jake Alexander",
    "Josh Moreira",
    "Lilly Roberts",
    "Matthew Curtis",
    "Naomi Schaufele",
    "Ryan Fang",
    "Sebastian Chen",
    "Tiansheng Sui",
    "Bryce Davidson"
]
maintainers = [
    "Ian Boyes <igboyes@gmail.com>"
]
description = "Viral infection diagnostics using next-generation sequencing"
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = "~3.12"
aiohttp = {extras = ["speedups"], version = "^3.11"}
aiojobs = "^1.0.0"
alembic = {extras = ["tz"], version = "^1.14"}
arrow = "^1.3.0"
asyncpg = "^0.30.0"
bcrypt = "^4.2.1"
biopython = "^1.84"
certifi = "^2024.8.30"
click = "^8.0.1"
dictdiffer = "^0.9.0"
email-validator = "^2.0.0"
Faker = "^12.3.3"
motor = "^3.6.0"
msal = "^1.16.0"
openfga-sdk = "^0.1.1"
openpyxl = "^3.1.5"
orjson = "^3.8.0"
piccolo-theme = "^0.24.0"
PyYAML = "^6.0.0"
PyJWT = "^2.9.0"
pytest = "^8.3.3"
semver = "^2.13.0"
sentry-sdk = "^2.18.0"
setuptools = "^75.3.0"
SQLAlchemy = "2.0.24"
structlog-sentry = "^2.1.0"
syrupy = "^4.5.0"
uvloop = "^0.21.0"
virtool-core = "^15.0.0"
visvalingamwyatt = "^0.1.4"
deprecated = "^1.2.15"
jsonref = "^1.1.0"

[tool.poetry.scripts]
virtool = "virtool.config.cli:entry"

[tool.poetry.group.dev.dependencies]
mypy = "^1.6.0"
pytest-aiohttp = "^1.0.5"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.14.0"
pytest-profiling = "^1.7.0"
pytest-regressions = "^2.2.0"
pytest-rerunfailures = "^10.2"
pytest-structlog = "^0.6"
pytest-xdist = "^3.3.1"
ruff = "^0.4.1"
snakeviz = "^2.2.0"
sphinx = "^7.0.1"
sphinx-autobuild = "^2021.3.14"
sphinx-autodoc-typehints = "^1.24.0"
sphinx-autofixture = "^0.4.0"
sphinx-toolbox = "^3.5.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
python_files = ["test_*.py", "rev_*.py"]
testpaths = ["tests", "assets/revisions"]

[tool.ruff]
target-version = "py312"
fix = true

[tool.ruff.lint]
select = ["ALL"]
ignore = ["ANN101", "D107", "EM102", "FBT001", "PLR0913", "RUF001"]

[tool.ruff.per-file-ignores]
"tests/**" = ["ANN201", "ANN202", "D205", "S101", "S106"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
